backend/requirements.txt
fastapi
uvicorn
motor
pymongo
httpx
google-generativeai
python-dotenv
pydantic[email]